Clean bin/ directory
