+++ app/bin/tools/backfill_search_fields.dart
@@ -8,8 +8,7 @@ import 'package:gcloud/db.dart';
 
 import 'package:pub_dartlang_org/backend.dart' show detectTypes;
 import 'package:pub_dartlang_org/models.dart';
-
-import 'tools_common.dart';
+import 'package:pub_dartlang_org/service_utils.dart';
 
 Future main(List<String> args) async {
   int updatedCount = 0;
