+++ app/bin/tools/uploader.dart
@@ -9,7 +9,7 @@ import 'package:gcloud/db.dart';
 
 import 'package:pub_dartlang_org/models.dart';
 
-import 'tools_common.dart';
+import 'package:pub_dartlang_org/service_utils.dart';
 
 Future main(List<String> arguments) async {
   if (arguments.length < 2 ||
